fix(bezier): close the undefined behaviour in A2.3's falling factorial
`evaluate_bezier_deriv_1d` accumulated `p (p-1) ... (p-k+1)` in a `std::int64_t`,
mirroring the oracle's numba `int64`. It overflows: at degree 30 from order 14, at
degree 61 from order 11, and both are reachable through
`Bezier.evaluate_derivatives`, which imposes no ceiling on `orders`. Numba wraps,
because it does not trap integer overflow. C++ signed overflow is undefined, so the
faithful-looking translation was worse than its original. Confirmed under
`-fsanitize=undefined`, which this repository's own `gcc-debug` preset enables:
`runtime error: signed integer overflow: 17 * 745747076954880000`.

The accumulator is now unsigned. Unsigned arithmetic is modular by definition and
the C++20 conversion back to `std::int64_t` is the two's-complement
reinterpretation, so this reproduces the oracle's wrapped value bit for bit while
being defined. Verified: the three wrapped values at degree 30, orders 13 to 15,
are identical before and after, and the sanitizer is silent.

Widening to `double` was the other candidate and it is wrong. It would have removed
the undefined behaviour and broken parity in the same stroke, since past 2^53 a
double rounds where an int64 wraps.

Eight parity cases now cover the region, all past the overflow point. Neither
backend's answer is correct there in any useful sense; what is under test is that
they are the same, and that the C++ reaches it by a defined route. The suite had
capped `n_deriv` at 4 and degree at 17, so nothing reached it.

+++ cpp/include/pantr/bezier/bezier.hpp
@@ -51,6 +51,7 @@
 
 #include <cmath>
 #include <cstddef>
+#include <cstdint>
 #include <span>
 #include <vector>
 
@@ -292,12 +293,25 @@ void evaluate_bezier_deriv_1d(span2d<const T> ctrl, std::span<const T> points, i
             }
         }
 
-        // The falling factorial is exact integer arithmetic in the oracle, and the
-        // promotion it forces against a float32 table is the one place this kernel
-        // leaves T's width.
-        std::int64_t fac = static_cast<std::int64_t>(degree);
+        // The falling factorial, and it is accumulated UNSIGNED on purpose.
+        //
+        // The oracle holds it in a numba `int64`, which wraps on overflow, and it
+        // does overflow: at degree 30 from `n_deriv = 14`, at degree 61 from 11,
+        // both reachable through `Bezier.evaluate_derivatives`. Signed overflow is
+        // undefined in C++ where numba's is merely wrong, so the natural
+        // translation is worse than its original -- confirmed under
+        // `-fsanitize=undefined`, which the repository's own `gcc-debug` preset
+        // enables.
+        //
+        // Unsigned arithmetic is modular by definition and the C++20 conversion
+        // back to `std::int64_t` is the two's-complement reinterpretation, so this
+        // reproduces the oracle's wrapped value bit for bit while being defined.
+        // Widening to `double` would have removed the undefined behaviour and
+        // broken parity instead, since past 2^53 a double rounds where an int64
+        // wraps.
+        std::uint64_t fac = static_cast<std::uint64_t>(degree);
         for (std::size_t k = 1; k <= num_derivs; ++k) {
-            const Acc fac_acc = Acc(static_cast<double>(fac));
+            const Acc fac_acc = Acc(static_cast<double>(static_cast<std::int64_t>(fac)));
             for (std::size_t j = 0; j < order; ++j) {
                 const Acc scaled = static_cast<Acc>(bd_at(k, j)) * fac_acc;
                 for (std::size_t r = 0; r < rank; ++r) {
@@ -306,7 +320,8 @@ void evaluate_bezier_deriv_1d(span2d<const T> ctrl, std::span<const T> points, i
                                        scaled * static_cast<Acc>(at(ctrl, j, r)));
                 }
             }
-            fac *= static_cast<std::int64_t>(degree) - static_cast<std::int64_t>(k);
+            fac *= static_cast<std::uint64_t>(static_cast<std::int64_t>(degree) -
+                                              static_cast<std::int64_t>(k));
         }
     }
 }
